+++ source/FileSystem/distortos-sources.cmake
@@ -11,3 +11,5 @@ target_sources(distortos PRIVATE
 		${CMAKE_CURRENT_LIST_DIR}/Directory.cpp
 		${CMAKE_CURRENT_LIST_DIR}/File.cpp
 		${CMAKE_CURRENT_LIST_DIR}/FileSystem.cpp)
+
+include(${CMAKE_CURRENT_LIST_DIR}/littlefs/distortos-sources.cmake)